+++ Network/API/Telegram/Bot/Object/Update/Message/Content/File/Audio.hs
@@ -8,13 +8,14 @@ import "base" Data.Functor ((<$>))
 import "base" Text.Show (Show)
 import "text" Data.Text (Text)
 
-import Network.API.Telegram.Bot.Object.Update.Message.Content.File.Special.URI (URI)
-import Network.API.Telegram.Bot.Object.Update.Message.Content.File.Special.Duration (Duration)
-import Network.API.Telegram.Bot.Object.Update.Message.Content.File.Special.Filesize (Filesize)
-import Network.API.Telegram.Bot.Object.Update.Message.Content.File.Special.MIME (MIME)
-import Network.API.Telegram.Bot.Property (Identifiable (Identificator, ident))
+import Network.API.Telegram.Bot.Object.Update.Message.Content.File.Special (Duration, Filesize, MIME, URI)
+import Network.API.Telegram.Bot.Property (Accessible (access), Identifiable (Identificator, ident))
 
 data Audio = Audio URI Duration (Maybe Text) (Maybe Text) (Maybe MIME) (Maybe Filesize) deriving Show
+
+instance Accessible Duration Audio where
+	access f (Audio uri duration performer title mime fs) =
+		(\duration' -> Audio uri duration' performer title mime fs) <$> f duration
 
 instance Identifiable Audio where
 	type Identificator Audio = URI