+++ Network/API/Telegram/Bot/Object/Update/Message/Content/File/Video.hs
@@ -7,15 +7,22 @@ import "base" Data.Function (($))
 import "base" Data.Functor ((<$>))
 import "base" Text.Show (Show)
 
-import Network.API.Telegram.Bot.Object.Update.Message.Content.File.Special.Duration (Duration)
-import Network.API.Telegram.Bot.Object.Update.Message.Content.File.Special.Filesize (Filesize)
-import Network.API.Telegram.Bot.Object.Update.Message.Content.File.Special.MIME (MIME)
-import Network.API.Telegram.Bot.Object.Update.Message.Content.File.Special.Height (Height)
-import Network.API.Telegram.Bot.Object.Update.Message.Content.File.Special.Width (Width)
-import Network.API.Telegram.Bot.Object.Update.Message.Content.File.Special.URI (URI)
-import Network.API.Telegram.Bot.Property (Identifiable (Identificator, ident))
+import Network.API.Telegram.Bot.Object.Update.Message.Content.File.Special (Duration, Filesize, MIME, Height, Width, URI)
+import Network.API.Telegram.Bot.Property (Accessible (access), Identifiable (Identificator, ident))
 
 data Video = Video URI Duration Height Width (Maybe MIME) (Maybe Filesize) deriving Show
+
+instance Accessible Duration Video where
+	access f (Video uri duration height width mime fs) =
+		(\duration' -> Video uri duration' height width mime fs) <$> f duration
+
+instance Accessible Height Video where
+	access f (Video uri duration height width mime fs) =
+		(\height' -> Video uri duration height' width mime fs) <$> f height
+
+instance Accessible Width Video where
+	access f (Video uri duration height width mime fs) =
+		(\width' -> Video uri duration height width' mime fs) <$> f width
 
 instance Identifiable Video where
 	type Identificator Video = URI