+++ Network/API/Telegram/Bot/Object/Update/Message/Content/File/Voice.hs
@@ -7,13 +7,14 @@ import "base" Data.Function (($))
 import "base" Data.Functor ((<$>))
 import "base" Text.Show (Show)
 
-import Network.API.Telegram.Bot.Object.Update.Message.Content.File.Special.Duration (Duration)
-import Network.API.Telegram.Bot.Object.Update.Message.Content.File.Special.Filesize (Filesize)
-import Network.API.Telegram.Bot.Object.Update.Message.Content.File.Special.MIME (MIME)
-import Network.API.Telegram.Bot.Object.Update.Message.Content.File.Special.URI (URI)
-import Network.API.Telegram.Bot.Property (Identifiable (Identificator, ident))
+import Network.API.Telegram.Bot.Object.Update.Message.Content.File.Special (Duration, Filesize, MIME, URI)
+import Network.API.Telegram.Bot.Property (Accessible (access), Identifiable (Identificator, ident))
 
 data Voice = Voice URI Duration (Maybe MIME) (Maybe Filesize) deriving Show
+
+instance Accessible Duration Voice where
+	access f (Voice uri duration mime fs) =
+		(\duration' -> Voice uri duration' mime fs) <$> f duration
 
 instance Identifiable Voice where
 	type Identificator Voice = URI