+++ Network/API/Telegram/Bot/Object/Update/Moving.hs
@@ -9,11 +9,16 @@ import "base" Data.Functor ((<$>))
 import "base" Text.Show (Show)
 
 import Network.API.Telegram.Bot.Object.Sender (Sender)
+import Network.API.Telegram.Bot.Property (Accessible (access))
 
 data Moving
 	= Gone Sender Group
 	| Joined [Sender] Group
 	deriving Show
+
+instance Accessible Group Moving where
+	access f (Gone sender group) = (\group' -> Gone sender group') <$> f group
+	access f (Joined senders group) = (\group' -> Joined senders group') <$> f group
 
 instance FromJSON Moving where
 	parseJSON = withObject "Moving" $ \v ->